+++ src/components/Common/Container/Container.module.css
@@ -1,0 +1,8 @@
+.container {
+  max-width: 120rem;
+  min-width: 37.5rem;
+  width: 100%;
+  margin: 4rem auto;
+  padding-left: 2.4rem;
+  padding-right: 2.4rem;
+}
