+++ src/components/ComparisonStatus/ComparisonList.js
@@ -6,7 +6,7 @@ import Pagination from '../Common/Pagination/Pagination';
 import noImageIcon from '../../assets/no-image.png';
 import { formatAmount } from '../../utils/formatAmount';
 import ComparisonHeader from './ComparisonHeader';
-import Warn from '../Warn';
+import Warn from '../Common/Warning/Warn';
 
 const MAX_ITEMS = 10;
